Create Webscraping using js
This code uses the cheerio library to parse the HTML response from the NSE website and extract the relevant trading data. It then uses the request library to make an HTTP request to the NSE website to retrieve the trading data for a particular symbol. The AWS library is used to save the data to an S3 bucket in AWS Cloud.

To run this code, you will need to install the required libraries by running npm install cheerio request aws-sdk. You will also need to set up your AWS credentials so that the code can access your S3 bucket.

I hope this helps! Let me know if you have any questions.